refactor(app): move slash menu controller + types to the editor layer (slashMenuController.ts)
